fix: corrige mapeamento SMALLINT e falta de transacao nos updates atomicos da cascata
Encontrados ao validar a aplicacao de ponta a ponta apos as mudancas de
hoje (pre-existentes, do trabalho da sessao anterior):

- ItemFilaJpaEntity mapeava prioridade_clinica (SMALLINT no banco) como
  Integer, e o schema-validation do Hibernate recusava a subir a aplicacao.
  Corrigido para Short.
- VagaJpaRepository.mudarStatusSeAtual e ConviteJpaRepository.atualizarStatusSeAtual
  sao @Modifying @Query sem @Transactional: Spring Data exige transacao
  explicita para queries de update customizadas (diferente do save(), que
  ja e transacional). Sem isso, todo o motor de cascata quebrava com
  TransactionRequiredException tanto no listener Kafka quanto nos
  endpoints de aceitar/recusar.

Cascata completa validada manualmente ponta a ponta apos a correcao:
cancelamento -> vaga.liberada -> convite enviado -> aceite -> vaga
OCUPADA -> paciente removido da fila.

+++ vaga-zero-api/src/main/java/br/com/fiap/vagazero/fila/infrastructure/persistencia/ItemFilaJpaEntity.java
@@ -28,7 +28,7 @@ public class ItemFilaJpaEntity {
     private LocalDateTime dataEntrada;
 
     @Column(name = "prioridade_clinica", nullable = false)
-    private Integer prioridadeClinica;
+    private Short prioridadeClinica;
 
     @Column(name = "aceita_chamado_imediato", nullable = false)
     private boolean aceitaChamadoImediato;
@@ -40,7 +40,7 @@ public class ItemFilaJpaEntity {
     }
 
     public ItemFilaJpaEntity(
-            Long id, Long pacienteId, String especialidade, LocalDateTime dataEntrada, Integer prioridadeClinica,
+            Long id, Long pacienteId, String especialidade, LocalDateTime dataEntrada, Short prioridadeClinica,
             boolean aceitaChamadoImediato, BigDecimal raioMaximoKm) {
         this.id = id;
         this.pacienteId = pacienteId;
@@ -67,7 +67,7 @@ public class ItemFilaJpaEntity {
         return dataEntrada;
     }
 
-    public Integer getPrioridadeClinica() {
+    public Short getPrioridadeClinica() {
         return prioridadeClinica;
     }
